support eventhub access with managed identity
