+++ DataProcessing/datax-host/pom.xml
@@ -170,11 +170,6 @@ SOFTWARE
             <artifactId>dbutils-api_2.11</artifactId>
             <version>0.0.3</version>
         </dependency>
-        <dependency>
-            <groupId>com.microsoft.azure</groupId>
-            <artifactId>msal4j</artifactId>
-            <version>1.12.0</version>
-        </dependency>
     </dependencies>
   <profiles>
     <profile>
