+++ DataProcessing/datax-keyvault/pom.xml
@@ -102,6 +102,11 @@ SOFTWARE
       <artifactId>jackson-databind</artifactId>
       <version>2.13.2.2</version>
     </dependency>
+    <dependency>
+      <groupId>com.microsoft.datax</groupId>
+      <artifactId>datax-core_2.4_2.11</artifactId>
+      <version>1.2.11-SNAPSHOT</version>
+    </dependency>
     <!--Testing-->
     <dependency>
       <groupId>org.jmockit</groupId>
